+++ src/main/java/asura/data/exception/AsuraException.java
@@ -1,3 +1,5 @@
+package asura.data.exception;
+
 public class AsuraException extends Exception{
     public AsuraException(String msg) {
         super(msg);
