+++ src/main/java/asura/data/tasks/Task.java
@@ -1,3 +1,5 @@
+package asura.data.tasks;
+
 public class Task {
     protected String description;
     protected boolean isDone;
@@ -10,6 +12,14 @@ public class Task {
 
     public String getStatusIcon() {
         return (isDone ? "[X]" : "[ ]");
+    }
+
+    public boolean getIsDone() {
+        return isDone;
+    }
+
+    public String getDescription() {
+        return description;
     }
 
     public void markAsDone() {
