+++ src/main/java/asura/ui/Ui.java
@@ -1,3 +1,5 @@
+package asura.ui;
+
 import java.util.Scanner;
 
 public class Ui {
