+++ text-ui-test/EXPECTED.TXT
@@ -1,5 +1,5 @@
    ---------------------------------------
-   Hello! I'm Asura!
+   Hello! I'm asura.Asura!
    What can I do for you?
    ---------------------------------------
